Refactor: Extract pageReducer tests from reducers.spec.js
Closes #204
